simplify token filter
